==> dataset_preprocess/4.filter_uniprot_dataset.ipynb <==
import os
import pandas as pd

# 加载从uniprot下载的酶数据集
uniprot_raw_path = 'ec_datasets/uniprot_raw'

download_uniprot_dataset = pd.read_csv(os.path.join(uniprot_raw_path, 'uniprot-download_sequence_site_ec.tsv'), sep='\t')
download_uniprot_dataset.head()

download_uniprot_dataset = download_uniprot_dataset.drop(['in_ec_react'], axis=1)
download_uniprot_dataset['Length'].describe()

# 为了节省计算资源，将最大的蛋白序列数限定在600个氨基酸
download_uniprot_dataset = download_uniprot_dataset.loc[download_uniprot_dataset['Length'] <= 600]
download_uniprot_dataset = download_uniprot_dataset.loc[download_uniprot_dataset['AlphaFoldDB'].notna()]
download_uniprot_dataset['Length'].describe()

download_uniprot_dataset.columns.tolist()

download_uniprot_dataset.info()

for col in [ 'Active site', 'Binding site', 'Site']:
    print(f'列 {col} 包含位点的数据：')
    print(len(download_uniprot_dataset.loc[~pd.isna(download_uniprot_dataset[col])]))

print('至少包含一种位点的数据：')
print(len(download_uniprot_dataset.loc[~(pd.isna(download_uniprot_dataset['Site']) & pd.isna(download_uniprot_dataset['Active site']) & pd.isna(download_uniprot_dataset['Binding site']))]))
    

def split_id(id_data):
    if pd.notna(id_data):
        return [x.strip() for x in id_data.split(';') if x]
    else:
        return id_data

# col_splits = ['EC number', 'PDB']
# for col in col_splits:
#     download_uniprot_dataset[col] = download_uniprot_dataset[col].apply(lambda x:split_id(x))

download_uniprot_dataset.head()

download_uniprot_dataset.info()

from collections import OrderedDict
def convert_site_data(site_data, split_flag):
    if pd.notna(site_data):
        site_data_list = [x.strip() for x in site_data.split(split_flag)]
        site_data_list = [x for x in site_data_list if x]
        reformate_site_data_list  = []
        for one_site in site_data_list:
            one_site_dict = OrderedDict()
            one_site_list = one_site.split('; /')
            for i, data in enumerate(one_site_list):
                if i == 0:
                    one_site_dict['function'] = split_flag
                    one_site_dict['position'] = data
                else:
                    info_name, info = data.split('=')
                    info = info.replace('\";', '')
                    info = info.replace('\"', '')
                    one_site_dict[info_name] = info
            reformate_site_data_list.append(one_site_dict)
        return reformate_site_data_list    
    else:
        return site_data

print('去重之后的序列数：')
print(len(set(download_uniprot_dataset['Sequence'])))

download_uniprot_dataset = download_uniprot_dataset.drop_duplicates(subset=['Active site', 'Binding site', 'Site', 'Sequence']).reset_index(drop=True)
download_uniprot_dataset.info()

download_uniprot_dataset.head()


# 重新整理活性位点的数据，用字典储存每一个位点
reformate_cols = ['Active site', 'Binding site', 'Site',]
split_flag_dict = {
    'Active site':'ACT_SITE', 'Binding site':'BINDING', 'Site':'SITE',
}
for col in reformate_cols:
    download_uniprot_dataset['{} reformate'.format(col)] = download_uniprot_dataset[col].apply(lambda x:convert_site_data(x, split_flag_dict[col]))

download_uniprot_dataset.head()

# 多个EC number的数据复制，使得每个EC number都直接可以对应一组数据
download_uniprot_dataset['EC number'] = download_uniprot_dataset['EC number'].map(lambda x:x.split(';'))
download_uniprot_dataset = download_uniprot_dataset.explode('EC number')
download_uniprot_dataset['EC number'] = download_uniprot_dataset['EC number'].apply(lambda x:x.strip())

download_uniprot_dataset.head(10)

download_uniprot_dataset['Site labeled'] = ~(pd.isna(download_uniprot_dataset['Active site']) & pd.isna(download_uniprot_dataset['Site']) & pd.isna(download_uniprot_dataset['Binding site']))

site_nolabeled_dataset = download_uniprot_dataset.loc[~download_uniprot_dataset['Site labeled']]
site_nolabeled_dataset.info()    # 完全没有位点标记的数据

site_labeled_dataset = download_uniprot_dataset.loc[download_uniprot_dataset['Site labeled']]
site_labeled_dataset.info()    # 完全没有位点标记的数据

site_nolabeled_dataset.to_pickle(os.path.join(uniprot_raw_path, 'uniprot-download_sequence_site_ec_clean-site-nolabeled.pkl'))
site_labeled_dataset.to_pickle(os.path.join(uniprot_raw_path, 'uniprot-download_sequence_site_ec_clean-site-labeled.pkl'))
download_uniprot_dataset.to_pickle(os.path.join(uniprot_raw_path, 'uniprot-download_sequence_site_ec_clean.pkl'))
